+++ packages/swirl-components/src/components/swirl-resource-list/swirl-resource-list.tsx
@@ -15,15 +15,17 @@ export class SwirlResourceList {
 
   componentDidLoad() {
     this.collectItems();
-    this.removeItemsFromTabOrder();
   }
 
   private collectItems() {
     this.items = Array.from(
-      this.el.querySelectorAll(
+      this.el.querySelectorAll<HTMLSwirlResourceListItemElement>(
         "swirl-resource-list-item, swirl-resource-list-file-item"
       )
-    );
+    ).filter((el) => el.isConnected);
+
+    this.removeItemsFromTabOrder();
+    this.enableItemFocus(this.items[this.focusedIndex]);
   }
 
   private removeItemsFromTabOrder() {
@@ -36,28 +38,42 @@ export class SwirlResourceList {
     );
   }
 
+  private enableItemFocus(
+    item?: HTMLSwirlResourceListItemElement,
+    focus?: boolean
+  ) {
+    if (!Boolean(item)) {
+      return;
+    }
+
+    const interactiveElement = item.shadowRoot?.querySelector<HTMLElement>(
+      ".resource-list-item__content, .resource-list-file-item"
+    );
+
+    if (!Boolean(interactiveElement)) {
+      return;
+    }
+
+    interactiveElement.setAttribute("tabIndex", "0");
+
+    if (focus) {
+      interactiveElement.focus();
+    }
+  }
+
   private focusItemAtIndex(index: number) {
     this.removeItemsFromTabOrder();
 
     const item = this.items[index];
 
-    if (!Boolean(item)) {
+    if (!Boolean(item) || !item.isConnected) {
       return;
     }
 
-    const interactiveElement = item.shadowRoot.querySelector<HTMLElement>(
-      ".resource-list-item__content, .resource-list-file-item"
-    );
-
-    interactiveElement.setAttribute("tabIndex", "0");
-    interactiveElement.focus();
+    this.enableItemFocus(item, true);
 
     this.focusedIndex = index;
   }
-
-  private onFocus = () => {
-    this.focusItemAtIndex(this.focusedIndex);
-  };
 
   private onKeyDown = (event: KeyboardEvent) => {
     if (event.key === "ArrowDown") {
@@ -81,16 +97,15 @@ export class SwirlResourceList {
     }
   };
 
+  private onSlotChange = () => {
+    this.collectItems();
+  };
+
   render() {
     return (
       <Host onKeyDown={this.onKeyDown}>
-        <swirl-stack
-          aria-label={this.label}
-          onFocus={this.onFocus}
-          role="grid"
-          tabIndex={0}
-        >
-          <slot></slot>
+        <swirl-stack aria-label={this.label} role="grid">
+          <slot onSlotchange={this.onSlotChange}></slot>
         </swirl-stack>
       </Host>
     );