+++ packages/swirl-tokens/dart/lib/styles.dark.dart
@@ -4,7 +4,7 @@
 //
 
 // Do not edit directly
-// Generated on Fri, 03 Mar 2023 10:53:44 GMT
+// Generated on Fri, 03 Mar 2023 12:30:43 GMT