+++ packages/swirl-components/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @getflip/swirl-components
 
+## 0.24.2
+
+### Patch Changes
+
+- [`eab7cec`](https://github.com/getflip/swirl/commit/eab7cec006f01043a5b24a308892d045fb7bfdcd)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix swirl-resource-list focus
+  behavior in Safari
+
+## 0.24.1
+
+### Patch Changes
+
+- [`230ecae`](https://github.com/getflip/swirl/commit/230ecae058fedcf0a5845c70c1826ed5413f9999)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix swirl-resource-list focus
+  handling
+
+## 0.24.0
+
+### Minor Changes
+
+- [`328e9f7`](https://github.com/getflip/swirl/commit/328e9f7ee1db98e54c897a67a5bf3317931d5a2a)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "custom-header" slot to
+  swirl-modal
+
+## 0.23.1
+
+### Patch Changes
+
+- [`c70da87`](https://github.com/getflip/swirl/commit/c70da8777f9cab93e8146edbf7bdda418a4df586)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix focus handling for updating
+  swirl-resource-list
+
+## 0.23.0
+
+### Minor Changes
+
+- [#178](https://github.com/getflip/swirl/pull/178)
+  [`ce4686a`](https://github.com/getflip/swirl/commit/ce4686a40c1fa3c21d4d3d58b15b7ab4d6d11405)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "max-width" prop to
+  swirl-modal
+
+* [#178](https://github.com/getflip/swirl/pull/178)
+  [`ce4686a`](https://github.com/getflip/swirl/commit/ce4686a40c1fa3c21d4d3d58b15b7ab4d6d11405)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add swirl-app-bar component
+
+- [#178](https://github.com/getflip/swirl/pull/178)
+  [`ce4686a`](https://github.com/getflip/swirl/commit/ce4686a40c1fa3c21d4d3d58b15b7ab4d6d11405)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add "truncate" prop to
+  swirl-heading
+
 ## 0.22.0
 
 ### Minor Changes